No telp dan benerin OTP regist

+++ database/migrations/2025_09_02_143853_create_user_questionnaire_answers_table.php
@@ -8,13 +8,15 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('user_questionnaire_answers', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->string('question_key');
-            $table->text('answer_value');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('user_questionnaire_answers')) {
+            Schema::create('user_questionnaire_answers', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->string('question_key');
+                $table->text('answer_value');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
